refactor: added additional sub folders to make domain folder more descriptive
