+++ src/enemy/store/enemies.ts
@@ -1,4 +1,4 @@
-import { Enemy } from "./enemy";
+import { Enemy } from "../enemy";
 
 export class Enemies {
     protected data: Enemy[] = []
