+++ src/tower/services/addTower.test.js
@@ -5,12 +5,12 @@
 import { describe, expect, test, beforeEach } from '@jest/globals';
 import { Tower } from '../entities/tower';
 import { AddTower } from './addTower';
-import { Grid } from '../placements/grid';
+import { Grid } from '../../grid/grid';
 import { Towers } from './towers';
-import { renderGrid } from '../placements/renderGrid';
-import { generateGrid } from '../helpers/grid';
-import { InvalidLocationError } from '../exceptions/InvalidLocationError';
-import { PathPlacement } from '../placements/pathPlacement';
+import { renderGrid } from '../../grid/services/renderGrid';
+import { generateGrid } from '../../helpers/grid';
+import { InvalidLocationError } from '../../exceptions/InvalidLocationError';
+import { PathPlacement } from '../../grid/placement/pathPlacement';
 
 
 describe('add tower service', () => {
